rename domain to common
