+++ docker-compose.ci.yml
@@ -14,6 +14,7 @@ services:
     ports:
       - 8000:8000
     container_name: riksutin_test
+    command: npm run start:test
 
   db:
     image: postgres:15.7
