+++ src/client/components/Common/SelectFaculty.tsx
@@ -11,7 +11,7 @@ import useUserFaculties from '../../hooks/useUserFaculties'
 
 import ShowMore from './ShowMore'
 
-import { extraOrganisations, organisationInfos } from '@domain/organisations'
+import { extraOrganisations, organisationInfos } from '@common/organisations'
 
 import styles from '../../styles'
