+++ src/common/getRiskTexts.ts
